manual merge from 'ocean-usermanagement'

+++ restoauth-webapp-oceanbank/src/main/java/com/oceanbank/webapp/restoauth/dao/UserRepository.java
@@ -24,7 +24,7 @@ import com.oceanbank.webapp.restoauth.model.DashboardUser;
  * @since 03.10.2015
  */
 @Transactional(readOnly = true)
-public interface UserRepository extends JpaRepository<DashboardUser, Long> {
+public interface UserRepository extends JpaRepository<DashboardUser, Integer> {
 	
 	/**
 	 * Find by user id is.
